requirements.txt
pandas~=2.2.3
matplotlib~=3.10.0
pillow~=11.1.0